refactor(gfql): move the var-length row specializations out of row_pipeline.py
Pure code move, no behaviour change. row_pipeline.py 1638 -> 1528 lines; the
unbounded/variable-length arms live in polars/varlen_rows.py with a docstring
explaining the exhaustion-depth walk and why deduping by endpoint is sound.

Motivation is rebase surface: #1757, #1714 and the seeded-chain work all touch
row_pipeline.py, and ~270 lines of var-length specialization sitting in the core
flow made every one of those a conflict.

RowPipelineMixin stays imported FUNCTION-LOCALLY exactly as at the original site —
hoisting it to module scope reintroduces an import cycle (row.pipeline -> lazy
engine -> back), so the move preserves it and says why.

2244 polars tests pass, ruff clean.
